arm: Populate bitops
Currently the bit operations are not populated for arm, thus implement
them in both atomic and non-atomic (with a _ prefix) versions.
Atomic operations are protected by the calls to new macros:
local_irq_save and local_irq_restore.

This patch acts as a prerequisite before mapping the shared info.
It is also necessary to define shared_info in this patch. Otherwise
we would get a compiler error as there are already common functions
passing shared_info to bitops.

Signed-off-by: Michal Orzel <michal.orzel@arm.com>
Change-Id: I1022583215d6efd0b50f76a255dd4962e07fd6ef