arm: Add support for handling virtual timer
Implement all the necessary functions/helpers to make use of a virtual
timer functionality. Modify the GIC IRQ handlers to call a virtual
timer interrupt handler if the interrupt ID is 27.

Virtual timer interrupt handler is marked as __weak so that it can
be redefined by tests to suite their needs.

Signed-off-by: Michal Orzel <michal.orzel@arm.com>
Change-Id: Ifbfc48c956eb48288d6e5a8568ee22735454f749