mm: Map pv_console ring buffer as memory
Introduce a new mapping flag DESC_PAGE_TABLE_MEM to be used to map some
area as normal memory.
Fix pv_console setup to map the console ring as memory and not as
device.
This is solving issues with console getting stuck because xtf does not
see memory changes done by the other side due to wrong coherency.

Signed-off-by: Bertrand Marquis <bertrand.marquis@arm.com>
Change-Id: I21675901f2c306f966fbaff67d025d8a24e23bee