arm64: Move the declaration of _start symbol before Image header
XTF_VIRT_START is defined to be 0xFFFFFF8000000000.
However, because the global symbol _start is declared after
the definition of the Image header, the entry point address stored in
the ELF header is incorrect as it reports XTF_VIRT_START + 64 bytes.
Also the effective image size stored in the Image header is wrong
as it does not include the header itself.

Fix this by moving the symbol _start right before the Image header
definition. Branch to kernel_start function instead of _start.

Signed-off-by: Michal Orzel <michalorzel.eng@gmail.com>