arm: Harden setup_pv_console
Currently setup_pv_console is called when XTF is used as domU. However
PV console is not available when running XTF as dom0less domU. In that
case the uninitialized hvm parameters are 0 and no error is returned.
Take this into account by adding a check to see if the pfn is 0.
If it is, skip PV console initialization.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>