build: Allow tests for different architectures
It might be that some tests are intended only for some specific
architectures. To handle this scenario introduce TEST-ARCH variable
that can be set in a test's Makefile to specify the architectures.
If not set at all, it defaults to all supported architectures. Skip the
test if the architecture we are building the test for is not listed in
TEST-ARCH.

It might be that TEST-ENVS contains environments for different
architectures (e.g. TEST-ENVS := mmu64le pv64). In that case skip the
unrecognized environments and override TEST-ENVS to contain only the
recognized ones. Skip the test if the TEST-ENVS is empty at this point.

Introduce SKIP_TEST variable in gen.mk, that if set, will define empty
build and install targets to skip processing the test.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>