arm: Add simple Arm SBSA UART driver
At the moment, we use Xen hypercall console to print messages. However,
on Xen production builds this console is disabled. In such cases, when
running as domU, we can make use of virtual PL011 UART that Xen
exposes to guests. Add a driver for this device (Xen implements vpl011
as SBSA UART) so that we can print messages when running as domU on Xen
production builds.

Introduce config option CONFIG_SBSA_UART, that if set, will enable
SBSA UART driver and use it to drive vpl011 device exposed by Xen.

When running as dom0less direct-mapped domU, the address of vpl011
is the address of Xen dtuart (otherwise it is using virtual Xen PL011
mapping). Introduce CONFIG_SBSA_UART_ADDRESS for that purpose.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>